update ft_free(): freed imgs
